\name{bignmf-package}
\alias{bignmf-package}
\docType{package}
\title{
Solving NMF via coordinate descent
}
\description{
The bignmf package solves the nonnegative matrix factorization via alternating least square. 
It treats nonnegative constrained regression as a special L1 regression and solves it via 
coordinate descent method which is quite fast. The core code is implemented in C++ using RcppEigen.
}
\details{
\tabular{ll}{
Package: \tab bignmf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-24\cr
License: \tab GPL \cr
}
}
\author{
Lanfeng Pan, Yixuan Qiu and Taiyun Wei
Maintainer: Lanfeng Pan <panlanfeng@gmail.com>
}
% \references{
% ~~ Literature or other references for background information ~~
% }


\keyword{optimize}

